requirements.txt
flask-sqlalchemy
lxml
pip-chill
pymysql
python-dotenv
